Backend setup complete

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @desc Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-	const goals = await Goal.find();
+	const goals = await Goal.find({ user: req.user.id });
 
 	res.status(200).json(goals);
 });
@@ -15,7 +16,6 @@ const getGoals = asyncHandler(async (req, res) => {
 // @route set /api/goals
 // @access Private
 const setGoal = asyncHandler(async (req, res) => {
-	console.log(req.body);
 	if (!req.body.text) {
 		res.status(400);
 		throw new Error('Please add in new text field');
@@ -23,6 +23,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
 	const goal = await Goal.create({
 		text: req.body.text,
+		user: req.user.id,
 	});
 
 	res.status(200).json(goal);
@@ -37,6 +38,18 @@ const updateGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400);
 		throw new Error('Goal not found');
+	}
+
+	const user = await User.findById(req.user.id);
+
+	if (!user) {
+		res.status(401);
+		throw new Error('User not found');
+	}
+
+	if (goal.user.toString() !== user.id) {
+		res.status(401);
+		throw new Error('User not authorized');
 	}
 
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -55,6 +68,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400);
 		throw new Error('Goal not found');
+	}
+
+	const user = await User.findById(req.user.id);
+
+	if (!user) {
+		res.status(401);
+		throw new Error('User not found');
+	}
+
+	if (goal.user.toString() !== user.id) {
+		res.status(401);
+		throw new Error('User not authorized');
 	}
 
 	const deletedGoal = await Goal.findByIdAndDelete(req.params.id, req.body);
